Rename internal modules so they don't clash with packages
